requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-allauth==0.44.0
django-extensions==3.1.3
django-heroku==0.3.1
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.8.1
Pillow==8.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pylint==2.8.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
validate-email==1.3
Werkzeug==2.0.1
whitenoise==5.2.0
wrapt==1.12.1
